+++ fsleyes/views/scene3dpanel.py
@@ -89,6 +89,7 @@ class Scene3DPanel(canvaspanel.CanvasPanel):
         opts.bindProps('zoom',         sceneOpts)
         opts.bindProps('offset',       sceneOpts)
         opts.bindProps('rotation',     sceneOpts)
+        opts.bindProps('highDpi',      sceneOpts)
 
         sizer = wx.BoxSizer(wx.HORIZONTAL)
         sizer.Add(self.__canvas, flag=wx.EXPAND, proportion=1)